Moved advertiser and settings services to proper folders
