+++ src/app/app.module.ts
@@ -12,8 +12,8 @@ import { SettingsModule } from './settings/settings.module'
 import { AppComponent } from './app.component';
 import { environment } from '../environments/environment';
 
-import { AdvertiserService } from './store/advertiser/advertiser.service';
-import { SettingsService } from './store/settings/settings.service';
+import { AdvertiserService } from './advertiser/advertiser.service';
+import { SettingsService } from './settings/settings.service';
 
 import { EffectsModule } from '@ngrx/effects';
 import { AdvertiserEffects } from './store/advertiser/advertiser.effects';
