+++ src/app/settings/settings.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Http, Response } from '@angular/http';
 import 'rxjs/add/operator/map';
-import { environment } from '../../../environments/environment';
+import { environment } from '../../environments/environment';
 
 
 @Injectable()
